Replace gosu with su-exec
su-exec is a newer, smaller (10k vs 1.8MiB) binary available as an Alpine package. This eliminates the need to use GPG key verification. This has been a big problem because recv-keys keeps on giving errors/timeouts even with a variety of key servers, causing Docker builds to fail. Since gosu was the only thing we needed GPG key verification for, removing this dependency was the ideal solution.